requirements.txt
python-telegram-bot[job-queue]==21.6
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pytz==2024.1
